+++ cetlvast/suites/compile/test_unbounded_variant_footprint_get_const.cpp
@@ -1,5 +1,5 @@
 /// @file
-/// Compile test that ensures it's impossible set "bigger" value than `Footprint` of `any`.
+/// Compile test that ensures it's impossible get "bigger" value than `Footprint` of const `unbounded_variant`.
 ///
 /// @copyright
 /// Copyright (C) OpenCyphal Development Team  <opencyphal.org>
@@ -7,7 +7,7 @@
 /// SPDX-License-Identifier: MIT
 ///
 
-#include "cetl/any.hpp"
+#include "cetl/unbounded_variant.hpp"
 
 #include <cstdint>
 
@@ -25,21 +25,20 @@ int main()
 {
     using any = cetl::any<sizeof(uint8_t)>;
 
-    any test{};
+    const any test{static_cast<uint8_t>(0)};
 
 #ifndef CETLVAST_COMPILETEST_PRECHECK
 
-    // Verify at `cetl::detail::base_storage::make_handlers`
+    // Verify at `cetl::detail::base_storage::get_ptr const`
     // ```
-    // static_assert(sizeof(Tp) <= Footprint, "Enlarge the footprint");
+    // static_assert(sizeof(ValueType) <= Footprint,
+    //               "Cannot contain the requested type since the footprint is too small");
     // ```
-    test = static_cast<uint16_t>(1);
+    return cetl::any_cast<uint16_t>(test);
 
 #else
 
-    test = static_cast<uint8_t>(1);
+    return cetl::any_cast<uint8_t>(test);
 
 #endif
-
-    return 0;
 }
