+++ cetlvast/suites/docs/examples/example_10_unbounded_variant.cpp
@@ -1,12 +1,12 @@
 /// @file
-/// Example of using cetl::any.
+/// Example of using cetl::unbounded_variant.
 ///
 /// @copyright
 /// Copyright (C) OpenCyphal Development Team  <opencyphal.org>
 /// Copyright Amazon.com Inc. or its affiliates.
 /// SPDX-License-Identifier: MIT
 ///
-#include "cetl/any.hpp"
+#include "cetl/unbounded_variant.hpp"
 
 #include <gtest/gtest.h>
