+++ cetlvast/suites/unittest/test_unbounded_variant.cpp
@@ -1,12 +1,12 @@
 /// @file
-/// Unit tests for cetl/any.hpp
+/// Unit tests for cetl/unbounded_variant.hpp
 ///
 /// @copyright
 /// Copyright (C) OpenCyphal Development Team  <opencyphal.org>
 /// Copyright Amazon.com Inc. or its affiliates.
 /// SPDX-License-Identifier: MIT
 
-#include <cetl/any.hpp>
+#include <cetl/unbounded_variant.hpp>
 
 #include <complex>
 #include <functional>
